requirements.txt
mistralai==0.4.2
python-dotenv
pandas
matplotlib
streamlit